Adds a basic authentication module that adds a real identity to the logs.
Works only when a .aur file is used, since it requires code executed by a
base file to hook the authentication process. In the medium term, we need
to create our own client so that we can hook auth + updates for all methods
without going through .aur files that may or may not be imported.